New Gvec_FFT_distribution class.
Now wave-functions can be swapped differently for Hloc and rho.

+++ src/mpi_grid.h
@@ -62,13 +62,13 @@ class MPI_grid
             entire grid. */
         std::vector<Communicator> communicators_;
 
-        /// number of MPI ranks in each communicator
+        /// Number of MPI ranks in each communicator.
         std::vector<int> communicator_size_;
 
         /// true if this is the root of the communicator group
         std::vector<bool> communicator_root_;
 
-        /// return valid directions for the current grid dimensionality
+        /// Return valid directions for the current grid dimensionality.
         inline int valid_directions(int directions__) const
         {
             return (directions__ & ((1 << dimensions_.size()) - 1));
@@ -82,7 +82,7 @@ class MPI_grid
             if (dimensions_.size() == 0) TERMINATE("no dimensions provided for the MPI grid");
 
             int sz = 1;
-            for (int i = 0; i < (int)dimensions_.size(); i++) sz *= dimensions_[i];
+            for (size_t i = 0; i < dimensions_.size(); i++) sz *= dimensions_[i];
             
             if (parent_communicator_.size() != sz)
             {
